# packages/shared-types/openapi.yaml
openapi: 3.1.0
info:
  title: Luvin API
  version: 0.0.0
servers:
  - url: /v1
paths:
...
  /auth/register:
    post:
      summary: Register an account
      operationId: postAuthRegister
      responses:
        "201":
          description: Registration succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        "422":
          description: Validation, age, legal, or captcha failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
...
  /auth/sessions/{sessionId}:
    delete:
      summary: Revoke a session
      operationId: deleteAuthSession
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Session revoked
...
components:
  schemas:
    SuccessMeta:
      type: object
      required: [requestId]
      properties:
        requestId:
          type: string
...
    ApiErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, fields, requestId]
          properties:
            code:
              type: string
            message:
              type: string
            fields:
              type: object
              additionalProperties:
                type: string
            requestId:
              type: string
    AuthSessionData:
      type: object
      required:
        [
          accessToken,
          refreshToken,
          userId,
          sessionId,
          accountState,
          requiredLegalActions,
        ]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        userId:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        accountState:
          type: string
        requiredLegalActions:
          type: array
          items:
            type: string
            enum: [ACCEPT_TERMS, ACCEPT_PRIVACY]
    AuthSessionResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/AuthSessionData"
        meta:
          $ref: "#/components/schemas/SuccessMeta"